improve maestro documentation
A Maestro USB servo controller that is reset to default settings has "UART, detect baud rate" as the serial setting, which will not work with this framework's USB control. This commit changes the documentation to mention setting the Maestro to USB Dual Port mode and links to the user's guide for help with installing the drivers and control software.