Update game_mode.rst
I am suggesting some further information to clarify the second example, as I had to spend time trying to figure out what some of the items in the example meant.

"The example below creates a new mode called JAM_rollover, and uses a machine-wide player variable named pf_multiplier.  This variable is what can be used in other parts of the game logic to multiply values based on the current multiplier value, for example, when calculating end of ball bonuses.  The counter value lb_JAM_complete_count is used as the count value in the JAM_lanes_done{count==2} within the variable_player conditional event statements."