add debug: yes
NOTE: I am not sure if the way I wrote that was correct since I am going through the tutorial for the first time but the sentence says "For example, if you have a modern style trough with a jam switch, you'd add the debug setting like this:" and then doesn't show debug: yes in the code block? So I think that needs to be added? So I just guessed that it should come after the "eject_coil_jam_pulse: 15ms" line?