Fixed underline
This is the first time I've ever had to edit a pull request so I don't know if I'm doing it correctly but looks like there was an underline length error so I've added another dash to extend the underline to the end of the title section's name. Hope that was done correctly.