Update Complete_Events for clarity RE state order
Added notes starting at 216 to clarify that the achievement must be in started state to move to completed state.  I struggled with this for a couple hours because I couldn't get my start_events to post a correct show upon ball ending.  I removed that line, and it broke complete functionality.  Figured out that was it, and wanted to add to save someone issues down the road.