Updated State_Machine file
I added some additional detail to the difference between when the block is in a config file vs a mode file.  This tripped me up, and I think will be helpful.  Additionally added a related document as a skill shot tutorial.  I found this super useful in understanding how it actually works to implement it in my game.

This is my first one adding sections, so feel free to clean up any formatting if I screwed something up!  Thanks.