add note for toggle option
I think the toggle option needs more information for beginners. I have no hardware and realized it very late. If a beginner does not install the mpf-examples-dev folder with the MPF installation (like me) there are no further notes on toggel an that it is related to the keyboard section.