Update picture_in_picture.rst
Hi Jan, I saw that this doc had not been written yet so I thought I'd try to help out and write this up.  I have picture in picture working in Cuphead but if my technique or explanation is not correct, please feel free to delete this.  I don't know if I've got the format correct for linking to other docs but I copied what I saw in other docs so I hope it's right.