Add info on shutting down computer using MPF
Added an example of how you can use MPF to safely shutdown a computer. This has been incredibly helpful for me when developing on a Windows 10 PC when I don't have a monitor attached. 
Other people who are developing on a DMD-based game and don't have a monitor/keyboard/mouse attached may also find this useful.
Thanks to Wolfmarsh for the code.