Update leds.rst
The math is a bit off here I think (I just tested it on my FAST Nano). The last chain should be 192-255 (not 196-255). Additionally, while I know a lot of programming things start with "0" as the first group, the Nano circuit board and diagrams are printed as CH1, CH2, CH3, CH4. FAST also labels the headers as "Channels" not "Chains" but I think that could get confusing since the "Channels" section is right after this in the documentation (and is not referring to the 4 headers but the non-RGB LED type channels).