Adding note to vpx docs (windows only)
VPX is Visual Pinball X right? 
I was/am mildly unclear so please reject or correct me if I'm way off base, but the difference in word choice ie virtual vs visual made this a bit difficult to track down ...  but just trying to add some annotation that it is windows only software and outgoing link to explain what the third party software is.  

I think ideally going in and changing the platform from `platform: virtual_pinball` to clarify vpx in the code might help as well but that is a much larger change. Additionally I see that there are `virtual`, `smart_virtual` and `virtual_pinball` which are all the modes for running without physical devices, and they all have the virtual keyword which makes sense. Not trying to overstep, just adding something that would have helped me a day ago.