add balldevice_ball_device_ball_eject_attempt.md clone to events/queue_events/ as a hack
For whatever reason (probably ordering somewhere in mkdocs or even the host OS) this page is the only
one of the pages listed in the Queue Events duplicate reference that hijacks where the nav shows that you are.
The rest show you in their normal group even when clicked from the Queue Event sub listings.

The real fix here is for each of the queue event duplicate listings to get a second file, so that
inbound links can decide which context in which to show the event. Alternatively we could
make the queue event links all nested in queue_events/ but then make them all just redirects to the real locations.